LayoutTests/fast/dom/Window/window-custom-prototype-expected.txt

Test what happens when you set the window's prototype to various values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS this.__proto__ = window; this.__proto__ threw exception TypeError: cyclic __proto__ value.
PASS this.__proto__ = chainPointingBackToWindow; this.__proto__ threw exception TypeError: cyclic __proto__ value.
PASS this.__proto__ = 1; this.__proto__ is originalWindowPrototype
PASS this.__proto__ = 'a string'; this.__proto__ is originalWindowPrototype
PASS this.__proto__ = anotherObject; this.__proto__ is anotherObject
PASS anotherObject.__proto__ = window; this.__proto__ threw exception TypeError: cyclic __proto__ value.
PASS this.__proto__ = 1; this.__proto__ is anotherObject
PASS this.__proto__ = 'a string'; this.__proto__ is anotherObject
PASS this.__proto__ = anotherObject; this.__proto__ is anotherObject
PASS this.__proto__ = originalWindowPrototype; this.__proto__ is originalWindowPrototype
PASS anotherObject.__proto__ = window; anotherObject.__proto__ is window
PASS this.__proto__ = null; this.__proto__ is undefined
PASS successfullyParsed is true

TEST COMPLETE
